+++ CHANGELOG.md
@@ -11,11 +11,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Rent payment
     - New fields
         - Payment note
+- Document upload
+    - New document types
+        - Water meter reading        
 
 ### Changed
 
 - Rent payment
     - Use due date as default payment date
+- Document upload
+    - Set creation and subject date seperately
 
 ### Fixed